# core-lending-delinquency-collections-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-delinquency-collections
  description: Loan Delinquency & Collections Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-collections
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/collection-cases/{collectionCaseId}/escalations/{escalationId}:
    get:
      tags:
        - CollectionEscalation
      summary: Get an escalation record by ID
      operationId: getById_3
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: escalationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollectionEscalationDTO'
    put:
      tags:
        - CollectionEscalation
      summary: Update an escalation record
      operationId: update_3
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: escalationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionEscalationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollectionEscalationDTO'
    delete:
      tags:
        - CollectionEscalation
      summary: Delete an escalation record
      operationId: delete_3
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: escalationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/collection-cases/{collectionCaseId}/escalations:
    get:
      tags:
        - CollectionEscalation
      summary: List or search escalations for a collection case
      operationId: findAll_3
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.collectionCaseId
          in: query
          description: Exact filter for collection case id
          required: false
          schema:
            type: string
        - name: filters.escalationLevel
          in: query
          description: Exact filter for escalation level
          required: false
          schema:
            type: string
            enum:
              - LEVEL_1
              - LEVEL_2
              - LEVEL_3
              - LEGAL
              - EXTERNAL
        - name: filters.escalationReason
          in: query
          description: Exact filter for escalation reason
          required: false
          schema:
            type: string
            enum:
              - BROKEN_PROMISE
              - NO_CONTACT
              - HIGH_AMOUNT
              - FRAUD_SUSPECT
        - name: filters.escalationDate
          in: query
          description: Exact filter for escalation date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[escalationDate].from
          in: query
          description: Filter escalation date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[escalationDate].to
          in: query
          description: Filter escalation date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.notes
          in: query
          description: Exact filter for notes
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseCollectionEscalationDTO'
    post:
      tags:
        - CollectionEscalation
      summary: Create a new escalation
      operationId: create_3
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionEscalationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollectionEscalationDTO'
...
components:
  schemas:
...
    CollectionEscalationDTO:
      required:
        - collectionCaseId
        - escalationDate
        - escalationLevel
        - escalationReason
      type: object
      properties:
        collectionEscalationId:
          type: string
          format: uuid
          readOnly: true
        collectionCaseId:
          type: string
          format: uuid
        escalationLevel:
          type: string
          enum:
            - LEVEL_1
            - LEVEL_2
            - LEVEL_3
            - LEGAL
            - EXTERNAL
        escalationReason:
          type: string
          enum:
            - BROKEN_PROMISE
            - NO_CONTACT
            - HIGH_AMOUNT
            - FRAUD_SUSPECT
        escalationDate:
          type: string
          format: date-time
        notes:
          maxLength: 2000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseCollectionEscalationDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CollectionEscalationDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...